Fix tests for CI
